+++ src/Contracts/WeasyPrintFactory.php
@@ -24,6 +24,5 @@ interface WeasyPrintFactory
   public function stream(string $filename, array $headers = [], StreamMode $mode = StreamMode::INLINE): StreamedResponse;
   public function download(string $filename, array $headers = []): StreamedResponse;
   public function inline(string $filename, array $headers = []): StreamedResponse;
-  public function putFile(string $path, string|null $disk = null, array $options = []): bool;
   public function getData(): string;
 }